Données des élus

Un jeu de données contenant des informations sur les élus des départements Loire-Atlantique, Gers ainsi que de la commune de Faverelles dans le Loiret.

Usage

elus_sample

Format

Un data.frame avec 9 752 lignes et 16 colonnes :

Code.du.département

Code unique du département (ex : 44 pour Loire-Atlantique, 32 pour Gers).

Libellé.du.département

Nom du département (ex : "Loire-Atlantique", "Gers", "Loiret").

Code.de.la.collectivité.à.statut.particulier

Code identifiant une collectivité spéciale.

Libellé.de.la.collectivité.à.statut.particulier

Nom de la collectivité spéciale.

Code.de.la.commune

Code unique de la commune (ex : 44000 pour Nantes, 32000 pour Auch...).

Libellé.de.la.commune

Nom de la commune (ex : "Nantes", "Auch", "Faverelles").

Nom.de.l.élu

Nom de l’élu.

Prénom.de.l.élu

Prénom de l’élu.

Code.sexe

Sexe de l’élu (M = masculin, F = féminin).

Date.de.naissance

Date de naissance (au format "Caractère", à mettre en "Date").

Code.de.la.catégorie.socio.professionnelle

Code socio-professionnel.

Libellé.de.la.catégorie.socio.professionnelle

Libellé de la catégorie socio-professionnelle.

Date.de.début.du.mandat

Début du mandat (au format "Caractère", à mettre en "Date").

Libellé.de.la.fonction

Fonction exercée (ex : maire, conseiller municipal...).

Date.de.début.de.la.fonction

Début de la fonction (au format "Caractère", à mettre en "Date").

Code.nationalité

Code nationalité de l’élu.

Source

📂 Trouver les données sur data.gouv.fr

Examples

data(elus_sample)
head(elus_sample)
#>   Code.du.département Libellé.du.département
#> 1                  32                   Gers
#> 2                  32                   Gers
#> 3                  32                   Gers
#> 4                  32                   Gers
#> 5                  32                   Gers
#> 6                  32                   Gers
#>   Code.de.la.collectivité.à.statut.particulier
#> 1                                           NA
#> 2                                           NA
#> 3                                           NA
#> 4                                           NA
#> 5                                           NA
#> 6                                           NA
#>   Libellé.de.la.collectivité.à.statut.particulier Code.de.la.commune
#> 1                                              NA              32001
#> 2                                              NA              32001
#> 3                                              NA              32001
#> 4                                              NA              32001
#> 5                                              NA              32001
#> 6                                              NA              32001
#>   Libellé.de.la.commune Nom.de.l.élu Prénom.de.l.élu Code.sexe
#> 1                Aignan     AURENSAN          Michel         M
#> 2                Aignan    BARATAULT        Philippe         M
#> 3                Aignan     BARNADAS         Mathieu         M
#> 4                Aignan       CALLAC    Marie-Pierre         F
#> 5                Aignan       CHANUT          Michel         M
#> 6                Aignan         DALY       Geraldine         F
#>   Date.de.naissance Code.de.la.catégorie.socio.professionnelle
#> 1        09/07/1952                                         75
#> 2        08/12/1957                                         33
#> 3        07/06/1981                                         52
#> 4        16/11/1966                                         34
#> 5        14/09/1959                                         12
#> 6        27/01/1973                                         85
#>                                       Libellé.de.la.catégorie.socio.professionnelle
#> 1                                                 Ancienne profession intermédiaire
#> 2                                                     Cadre de la fonction publique
#> 3                         Employé civil et agent de service de la fonction publique
#> 4                                               Professeur, profession scientifique
#> 5                                              Agriculteur sur moyenne exploitation
#> 6 Personne diverse sans activité professionnelle de moins de 60 ans (sauf retraité)
#>   Date.de.début.du.mandat Libellé.de.la.fonction Date.de.début.de.la.fonction
#> 1              28/06/2020                                                    
#> 2              28/06/2020                                                    
#> 3              28/06/2020                                                    
#> 4              28/06/2020  3ème adjoint au Maire                   03/07/2020
#> 5              28/06/2020                                                    
#> 6              28/06/2020                                                    
#>   Code.nationalité
#> 1               FR
#> 2               FR
#> 3               FR
#> 4               FR
#> 5               FR
#> 6               IE