Générer un rapport à partir d'un fichier Quarto

Cette fonction permet de générer un rapport au format HTML ou PDF à partir d'un fichier Quarto (.qmd). Elle utilise les informations d'une commune et d'un département fournis en paramètres, ainsi qu'un dataframe optionnel contenant des données pour enrichir le rapport.

Usage

generer_rapport(commune, departement, output, df = elus_sample)

Arguments

commune

Un entier ou une chaîne de caractères représentant le code INSEE de la commune concernée.

departement

Un entier ou une chaîne de caractères représentant le code INSEE du département concerné.

output

Une chaîne de caractères spécifiant le chemin et le nom du fichier de sortie. Le format du fichier doit être précisé en ajoutant ".html" ou ".pdf" à la fin (par exemple, "rapport_nantes.html" ou "rapport_nantes.pdf").

df

(optionnel) Un dataframe contenant les données nécessaires pour générer le rapport. Par défaut, la fonction utilise `elus_sample`, qui est inclus dans le package.

Value

Aucun. La fonction génère un fichier de rapport au format HTML ou PDF dans le chemin spécifié par `output`.

Details

- Si le dataframe utilisé est elus_sample ou qu'aucun n'est précisé, les codes commune et département doivent correspondre à ceux présents dans elus_sample.

- Si un autre dataframe est utilisé, il doit respecter une structure minimale pour garantir le bon fonctionnement de la fonction. Voir les warnings en cas d'incompatibilité ou consulter la documentation sur le site du package.

- La fonction utilise quarto::quarto_render() pour générer le rapport au format spécifié.

Examples

# Générer un rapport pour Nantes (44109) et la Loire-Atlantique (44) en HTML
generer_rapport(44109, 44, "rapport_Nantes_et_Loire_Atlantique.html")
#> 
#> 
#> processing file: rapport.qmd
#> 1/13                  
#> 2/13 [unnamed-chunk-1]
#> 3/13                  
#> 4/13 [unnamed-chunk-2]
#> 5/13                  
#> 6/13 [setup]          
#> 7/13                  
#> 8/13 [unnamed-chunk-3]
#> 9/13                  
#> 10/13 [unnamed-chunk-4]
#> 11/13                  
#> 12/13 [unnamed-chunk-5]
#> 13/13                  
#> output file: rapport.knit.md
#> 
#> pandoc --output ..\..\..\..\..\..\Desktop\firstlibdjayan\docs\reference\rapport_Nantes_et_Loire_Atlantique.html
#>   to: html
#>   standalone: true
#>   section-divs: true
#>   html-math-method: mathjax
#>   wrap: none
#>   default-image-extension: png
#>   
#> metadata
#>   document-css: false
#>   link-citations: true
#>   date-format: long
#>   lang: en
#>   title: Rapport
#>   author: Djayan DAËRON
#>   
#> Output created: ..\..\..\..\..\..\Desktop\firstlibdjayan\docs\reference\rapport_Nantes_et_Loire_Atlantique.html
#> 
#> 
#> ✅ rapport_Nantes_et_Loire_Atlantique.html a été généré avec succès.

# Générer un rapport pour Nantes en PDF
generer_rapport(44109, 44, "rapport_Nantes_et_Loire_Atlantique.pdf")
#> 
#> 
#> processing file: rapport.qmd
#> 1/13                  
#> 2/13 [unnamed-chunk-1]
#> 3/13                  
#> 4/13 [unnamed-chunk-2]
#> 5/13                  
#> 6/13 [setup]          
#> 7/13                  
#> 8/13 [unnamed-chunk-3]
#> 9/13                  
#> 10/13 [unnamed-chunk-4]
#> 11/13                  
#> 12/13 [unnamed-chunk-5]
#> 
#> Sorry, but pdfcrop.exe did not succeed.
#> 
#> The log file hopefully contains the information to get MiKTeX going again:
#> 
#>   C:\Users\Djayan\AppData\Local\MiKTeX\miktex\log\pdfcrop.log
#> 
#> Sorry, but pdfcrop.exe did not succeed.
#> 
#> The log file hopefully contains the information to get MiKTeX going again:
#> 
#>   C:\Users\Djayan\AppData\Local\MiKTeX\miktex\log\pdfcrop.log
#> 13/13                  
#> output file: rapport.knit.md
#> 
#> pandoc --output rapport.tex
#>   to: latex
#>   standalone: true
#>   pdf-engine: xelatex
#>   variables:
#>     graphics: true
#>     tables: true
#>   default-image-extension: pdf
#>   
#> metadata
#>   documentclass: scrartcl
#>   classoption:
#>     - DIV=11
#>     - numbers=noendperiod
#>   papersize: letter
#>   header-includes:
#>     - \KOMAoption{captions}{tableheading}
#>   block-headings: true
#>   title: Rapport
#>   author: Djayan DAËRON
#>   
#> 
#> Rendering PDF
#> running xelatex - 1
#>   This is XeTeX, Version 3.141592653-2.6-0.999996 (TeX Live 2024) (preloaded format=xelatex)
#>    restricted \write18 enabled.
#>   entering extended mode
#>   
#> running xelatex - 2
#>   This is XeTeX, Version 3.141592653-2.6-0.999996 (TeX Live 2024) (preloaded format=xelatex)
#>    restricted \write18 enabled.
#>   entering extended mode
#>   
#> 
#> Output created: ..\..\..\..\..\..\Desktop\firstlibdjayan\docs\reference\rapport_Nantes_et_Loire_Atlantique.pdf
#> 
#> 
#> ✅ rapport_Nantes_et_Loire_Atlantique.pdf a été généré avec succès.

# Voir la documentation pour la structure attendue du dataframe
# https://djayan-d.github.io/firstlibdjayan/