Résumé des informations d'un département

Cette fonction génère un résumé des informations pour un département donné. Elle vérifie que l'objet passé en argument est bien de la classe departement, puis retourne une liste contenant le nom du département et des statistiques sur celui-ci, notamment le nombre de communes, le nombre d'élus, la distribution des âges des élus, ainsi que des informations sur les élus les plus âgés et les plus jeunes, ainsi que les communes les plus jeunes et les plus âgées.

Usage

# S3 method for class 'departement'
summary(x)

Arguments

x

Un objet de classe `departement` (un `data.frame` contenant les informations des élus et des communes).

Value

Une liste contenant plusieurs éléments :

nom_departement : Le nom du département

nombre_communes : Le nombre de communes dans le département

nombre_elus : Le nombre d'élus dans le département

distribution_age_elus : La distribution des âges des élus dans le département

elu_plus_age : Le nom et l'âge de l'élu le plus âgé

elu_plus_jeune : Le nom et l'âge de l'élu le plus jeune

commune_plus_jeune : La commune avec la moyenne d'âge la plus faible

commune_plus_agee : La commune avec la moyenne d'âge la plus élevée

See also

Examples

# Créer un exemple de jeu de données pour le département de Loire-Atlantique (code 44)
df_loire_atlantique <- data.frame(
  Code.de.la.commune = rep(44109, 5),
  Libellé.de.la.commune = rep("Nantes", 5),
  Libellé.du.département = rep("Loire-Atlantique", 5),
  Nom.de.l.élu = c("Dupont", "Dupuis", "Martin", "Lemoine", "Durand"),
  Prénom.de.l.élu = c("Jean", "Marie", "Paul", "Anne", "Pierre"),
  Date.de.naissance = as.Date(c("1970-01-01", "1980-05-12", "1965-08-22", "1992-10-30", "1950-12-05")),
  Code.de.la.catégorie.socio.professionnelle = c(1, 2, 3, 4, 5),
  stringsAsFactors = FALSE
)
df_loire_atlantique <- creer_departement(df_loire_atlantique)  # Ajouter la classe "departement"
#> Warning: ⚠️ Attention : Certaines colonnes essentielles sont absentes : Code.du.département, Libellé.de.la.fonction
#> Cela peut entraîner des erreurs dans certaines fonctions du package.
#> Error in creer_departement(df_loire_atlantique): "Code.du.département" %in% colnames(df) is not TRUE

# Appeler la fonction pour obtenir un résumé des informations du département de Loire-Atlantique
summary.departement(df_loire_atlantique)
#> Error in summary.departement(df_loire_atlantique): could not find function "summary.departement"