Générer un rapport à partir d'un fichier Quarto

Cette fonction génère un rapport au format PDF à partir d'un fichier Quarto (.qmd) en utilisant les paramètres de la commune et du département fournis.

Usage

generer_rapport(commune, departement, output, df = NULL)

Arguments

commune

Le code (ex : INSEE) associé à la commune pour laquelle le rapport sera généré.

departement

Le code (ex : INSEE) associé au département pour lequel le rapport sera généré.

output

Un caractère spécifiant le chemin du fichier de sortie (par exemple, "rapport_nantes.html").

Value

Aucun. La fonction génère un fichier de rapport au format spécifié (par défaut html).

Examples

generer_rapport(44109, 44, "rapport_Nantes_et_Loire_Atlantique.html")
#> 
#> 
#> processing file: rapport.qmd
#> 1/13                  
#> 2/13 [unnamed-chunk-1]
#> 3/13                  
#> 4/13 [unnamed-chunk-2]
#> trying URL 'https://cloud.r-project.org/bin/windows/contrib/4.3/xfun_0.51.zip'
#> Content type 'application/zip' length 585825 bytes (572 KB)
#> ==================================================
#> downloaded 572 KB
#> 
#> trying URL 'https://cloud.r-project.org/bin/windows/contrib/4.3/jsonlite_1.9.0.zip'
#> Content type 'application/zip' length 1105952 bytes (1.1 MB)
#> ==================================================
#> downloaded 1.1 MB
#> 
#> trying URL 'https://cloud.r-project.org/bin/windows/contrib/4.3/ps_1.9.0.zip'
#> Content type 'application/zip' length 652010 bytes (636 KB)
#> ==================================================
#> downloaded 636 KB
#> 
#> trying URL 'https://cloud.r-project.org/src/contrib/processx_3.8.6.tar.gz'
#> Content type 'application/x-gzip' length 165190 bytes (161 KB)
#> ==================================================
#> downloaded 161 KB
#> 
#> * installing *source* package 'processx' ...
#> ** package 'processx' successfully unpacked and MD5 sums checked
#> ERROR: cannot remove earlier installation, is it in use?
#> * removing 'C:/Users/Djayan/AppData/Local/R/win-library/4.3/processx'
#> * restoring previous 'C:/Users/Djayan/AppData/Local/R/win-library/4.3/processx'
#> Warning in file.copy(lp, dirname(pkgdir), recursive = TRUE, copy.date = TRUE) :
#>   problem copying C:\Users\Djayan\AppData\Local\R\win-library\4.3\00LOCK-processx\processx\libs\x64\processx.dll to C:\Users\Djayan\AppData\Local\R\win-library\4.3\processx\libs\x64\processx.dll: Permission denied
#> 
#> The downloaded source packages are in
#> 	'C:\Users\Djayan\AppData\Local\Temp\RtmpK2V3gD\downloaded_packages'
#> * installing *source* package 'firstlibdjayan' ...
#> ** using staged installation
#> ** R
#> ** data
#> *** moving datasets to lazyload DB
#> ** inst
#> ** byte-compile and prepare package for lazy loading
#> ** help
#> *** installing help indices
#> ** building package indices
#> ** installing vignettes
#> ** testing if installed package can be loaded from temporary location
#> ** testing if installed package can be loaded from final location
#> ** testing if installed package keeps a record of temporary installation path
#> * DONE (firstlibdjayan)
#> 5/13                  
#> 6/13 [setup]          
#> 7/13                  
#> 8/13 [unnamed-chunk-3]
#> 9/13                  
#> 10/13 [unnamed-chunk-4]
#> 11/13                  
#> 12/13 [unnamed-chunk-5]
#> 13/13                  
#> output file: rapport.knit.md
#> 
#> pandoc --output ..\..\..\..\..\..\Desktop\firstlibdjayan\docs\reference\rapport_Nantes_et_Loire_Atlantique.html
#>   to: html
#>   standalone: true
#>   section-divs: true
#>   html-math-method: mathjax
#>   wrap: none
#>   default-image-extension: png
#>   
#> metadata
#>   document-css: false
#>   link-citations: true
#>   date-format: long
#>   lang: en
#>   title: Rapport
#>   author: Djayan DAËRON
#>   
#> Output created: ..\..\..\..\..\..\Desktop\firstlibdjayan\docs\reference\rapport_Nantes_et_Loire_Atlantique.html
#> 
#> 
#> Le rapport a été généré avec succès.