Résumé des informations d'un département

Cette fonction génère un résumé des informations pour un département donné. Elle retourne une liste contenant des statistiques sur le département, notamment le nombre de communes, le nombre d'élus, la distribution des âges des élus, ainsi que des informations sur les élus les plus âgés et les plus jeunes, ainsi que les communes les plus jeunes et les plus âgées.

Usage

# S3 method for class 'departement'
summary(x)

Arguments

x

Un objet de classe `departement` (un `data.frame` contenant les informations des élus et des communes).

Value

Une liste contenant plusieurs éléments : - `nom_departement` : Le nom du département, - `nombre_communes` : Le nombre de communes dans le département, - `nombre_elus` : Le nombre d'élus dans le département, - `distribution_age_elus` : La distribution des âges des élus dans le département, - `elu_plus_age` : Le nom et l'âge de l'élu le plus âgé, - `elu_plus_jeune` : Le nom et l'âge de l'élu le plus jeune, - `commune_plus_jeune` : La commune avec la moyenne d'âge la plus faible, - `commune_plus_agee` : La commune avec la moyenne d'âge la plus élevée.